fix(framework): the contracts move out, and a queue entry must carry its identity
Review found the ratchet first: this file grew to 489 lines against a frozen
baseline of 459, and `scan-file-size.mjs` says what to do about that — "pre-existing
debt may shrink but never grow — split instead of extending." Raising the number
is the option the ratchet exists to refuse.

`IExecutionControllerCallbacks`, `ITurnOptions`, `IQueuedInput` and `TSubmitFn`
move to `interactive-session-execution-contracts.ts`. 489 -> 442, and the seam is
not arbitrary: those describe what a CALLER supplies and what a submission
CARRIES, the class describes what the controller DOES with them. The same split
this change already made for `turn-contracts.ts`. Every name is re-exported, so
no importer changes. Baseline rewritten in the same commit, which the ratchet
also asks for.

AN ENTRY WITHOUT AN IDENTITY IS NOW REFUSED AT `enqueue`. `turnId` is optional on
the type and every settle path opens `if (turnId === undefined) return;` — so an
entry built without one takes the whole queued path and its caller's `completed`
never settles. That defect SHIPPED in this change: the queue half was inert
because the id was not threaded onto the entry, and nothing failed. One
comparison converts the silent hang into a thrown error where the entry is built.

Three co-drive fixtures were the only place such an entry existed; they mint an
id now. A double that cannot be built the way the real thing is built is testing
something else.

The full fix — typing the optionality away across `ITurnOptions`, `IQueuedInput`
and three registry methods — is RUNTIME-006, filed rather than half-done inside a
behavioural PR.

And `completionOf`'s doc claimed a rejected orphan "can only mean the submission
already settled". Review showed a second way there: `resumeTurnId` is documented
as set only by the queue drain and nothing enforces it. Corrected — and the fact
that the registry cannot tell the two apart is itself the argument for RUNTIME-006
rather than for guessing.

RAN: framework 1319, file-size ratchet passed at the new baseline, tsc clean.
