+++ packages/agent-framework/src/interactive/__tests__/co-drive-e5.test.ts
@@ -23,8 +23,22 @@ function controller(): SessionExecutionController {
   return new SessionExecutionController({} as never, {} as never, { emit: vi.fn() } as never);
 }
 
+/**
+ * A queue entry, WITH the identity a real one always carries.
+ *
+ * These fixtures built entries without a `turnId`, which the queue now refuses at `enqueue` — an
+ * entry with no identity cannot be settled by any refusal, so its caller waits forever, and that
+ * defect shipped once in RUNTIME-003. The fixtures were the only place such an entry existed; a
+ * double that cannot be built the way the real thing is built tests something else.
+ */
+let mintedTurns = 0;
 function entry(input: string, driverId: string, wakeTaskId?: string): IQueuedInput {
-  return { input, options: { driverId, ...(wakeTaskId ? { wakeTaskId } : {}) } };
+  mintedTurns += 1;
+  return {
+    input,
+    turnId: `co-drive-turn-${mintedTurns}`,
+    options: { driverId, ...(wakeTaskId ? { wakeTaskId } : {}) },
+  };
 }
 
 describe('co-drive queue (REMOTE-014 TC-01)', () => {
