+++ packages/agent-framework/src/interactive/turn-settler-registry.ts
@@ -56,7 +56,13 @@ export class TurnSettlerRegistry {
    * The queue drain re-enters `submit` for an input that was accepted earlier, and its caller is
    * holding the promise from THAT acceptance. Handing back the same one is the whole point: minting
    * a second promise here would settle something nobody is waiting on. Returns a rejected promise
-   * for an id this controller does not know, which can only mean the submission already settled.
+   * for an id this controller does not know.
+   *
+   * "Already settled" is the ordinary way to get there and was once written here as the ONLY way.
+   * Review showed another: `resumeTurnId` is documented as set only by the queue drain, and nothing
+   * enforces that, so an internal caller passing an unregistered id reaches this same rejection.
+   * Both are refusals the caller can act on, and neither is a state this registry can distinguish —
+   * which is the argument for typing the optionality away rather than for guessing here.
    */
   completionOf(turnId: string): Promise<IExecutionResult> {
     const registered = this.settlers.get(turnId);
