+++ scripts/harness/file-size-baseline.json
@@ -25,7 +25,7 @@
   "packages/agent-framework/src/index.ts": 670,
   "packages/agent-framework/src/interactive/interactive-session-background-tracker.ts": 334,
   "packages/agent-framework/src/interactive/interactive-session-base.ts": 387,
-  "packages/agent-framework/src/interactive/interactive-session-execution-controller.ts": 459,
+  "packages/agent-framework/src/interactive/interactive-session-execution-controller.ts": 443,
   "packages/agent-framework/src/interactive/interactive-session-execution.ts": 313,
   "packages/agent-framework/src/interactive/interactive-session-history-tracker.ts": 355,
   "packages/agent-framework/src/interactive/interactive-session-skill-router.ts": 357,
